+++ CHANGELOG.md
@@ -5,6 +5,13 @@ Todos los cambios notables en este proyecto serán documentados en este archivo.
 El formato está basado en [Keep a Changelog](https://keepachangelog.com/es-ES/1.0.0/),
 y este proyecto adhiere a [Semantic Versioning](https://semver.org/lang/es/).
 
+## [0.5.0] - 2025-10-23
+
+### Añadido
+- Función `calcularPromedio(numeros)` para obtener el promedio de una lista de números.
+- Opción para "Promedio" en el menú interactivo.
+- Tests unitarios para la función `calcularPromedio`.
+
 ## [0.4.0] - 2025-10-22
 
 ### Añadido